requirements.txt
fastapi
uvicorn
pandas
python-dotenv
twilio
openai
httpx
requests
